# Remove unnecessary defer in Span.End, shaving significant time