@rakyll 2018-03-15
Nice, I almost about to believe defer got so much better in the last release but we may actually go through a few things to avoid defer.