@odeke-em 2018-03-16
Offline @ramonza raised a great point that if ExportSpan happens to panic, this lock will be held forever since we don't have the defer. My response was that, yes I acknowledge that, but for the meantime, before we've encountered that problem let's tradeoff waiting for that day of reckoning with speed here.