@odeke-em 2018-03-16
@ramonza in the meantime I'll merge this PR, it is all internal stuff so we can revert if we feel strongly about it.