@rakyll 2018-03-16
If ExportSpan panics, the process crashes. I don't understand the last comment.