@odeke-em 2018-03-16
@rakyll, in the common case sure it will panic but `defer exportsMu.Unlock()` won't be invoked, in the case that Span.End() was invoked by a function that has recover() in it.