@rakyll 2018-03-16
In Go, we don't encourage recover at all. Don't optimize behavior for recover.