@odeke-em 2018-03-16
for example

```go
func pageContent(w http.ResponseWriter, r *http.Request) {
   defer func() {
       if r := recover(); r != nil {
             log.Printf("Handler crashed: %v\n", r)
       }
   }()

   ctx, span := trace.StartSpan(r.Context(), "/template")
   defer span.End()
   ...
}
```
I've seen a bunch of handlers with recovers in them. In this case exportersMu.Lock() will be held forever yet the program will proceed normally.