@rakyll 2018-03-16
This brings the other question of whether it is a good practice to call ExportSpan in the critical path at all.